Make the `to-markdown` code handle newlines paragraph spacing in export
Blog posts don't wrap paragraphs in <p> tags.  As a result, a body that
includes successive paragraphs ends up with markdown that is a single
large paragraph.

FIX: The simplest fix seemed to be finding blank lines in the export
and making sure they stay blank lines in the markdown. The most
straight forward way to do this seemed to be exploiting the way
Turndown handles <br/> tags.  Turndown converts these to a newline.
Replacing blank lines with double <br/> elements seems to work as desired.